+++ net-snmp/agent/mibgroup/agent/nsDebug.c
@@ -291,7 +291,7 @@ get_first_debug_entry(void **loop_context, void **data_context,
 
     snmp_set_var_value(index, dbg_tokens[i].token_name,
 		       strlen(dbg_tokens[i].token_name));
-    *loop_context = (void*)i;
+    *loop_context = (void*)(intptr_t)i;
     *data_context = (void*)&dbg_tokens[i];
     return index;
 }
@@ -301,7 +301,7 @@ get_next_debug_entry(void **loop_context, void **data_context,
                       netsnmp_variable_list *index,
                       netsnmp_iterator_info *data)
 {
-    int i = (int)*loop_context;
+    int i = (int)(intptr_t)*loop_context;
 
     for (i++; i<debug_num_tokens; i++) {
         /* skip excluded til mib is updated */
@@ -313,7 +313,7 @@ get_next_debug_entry(void **loop_context, void **data_context,
 
     snmp_set_var_value(index, dbg_tokens[i].token_name,
 		       strlen(dbg_tokens[i].token_name));
-    *loop_context = (void*)i;
+    *loop_context = (void*)(intptr_t)i;
     *data_context = (void*)&dbg_tokens[i];
     return index;
 }
